requirements.txt
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
